+++ js/api.js
@@ -1,8 +1,10 @@
-const getData = (onSuccess) => {
+const getData = (onSuccess, onFail) => {
   fetch('https://27.javascript.pages.academy/kekstagram/data')
     .then((response) => response.json())
     .then((photos) => {
       onSuccess(photos);
+    }).catch(() => {
+      onFail();
     });
 };
