+++ js/forms.js
@@ -1,6 +1,7 @@
 import {isEscapeKey} from './util.js';
-import {showAlert} from './util.js';
 import {sendData} from './api.js';
+import {showSuccessMessage, showErrorMessage} from './message.js';
+import {resetImage} from './slider.js';
 
 //находим форму загрузки и редактирования изображения
 const photoLoadingButton = document.querySelector('.img-upload__input');
@@ -25,7 +26,6 @@ const onLoadFormEscKeydown = (evt) => {
 
 //открытие формы загрузки изображения
 const openLoadForm = () => {
-  //evt.preventDefault();
   photoLoadingForm.classList.remove('hidden');
   document.body.classList.add('modal-open');
   document.addEventListener('keydown', onLoadFormEscKeydown);
@@ -103,6 +103,10 @@ pristine.addValidator(commentsInput, checkCommentsLength, 'Длина комме
 //отправка формы
 
 const submitButton = photoLoadingForm.querySelector('.img-upload__submit');
+//const sliderElement = document.querySelector('.effect-level__slider');
+const valueElement = document.querySelector('.effect-level__value');
+const scaleInput = document.querySelector('input[name="scale"]');
+//const imagePreview = document.querySelector('.img-upload__preview img');
 
 const blockSubmitButton = () => {
   submitButton.disabled = true;
@@ -112,6 +116,18 @@ const blockSubmitButton = () => {
 const unblockSubmitButton = () => {
   submitButton.disabled = false;
   submitButton.textContent = 'Опубликовать';
+};
+
+const resetForm = () => {
+  loadingForm.reset(); //сброс полей формы
+  valueElement.value = 50; //сброс уровня насыщенности
+  scaleInput.value = '100%'; //возврат к масштабу 100%
+  resetImage();
+  // sliderElement.noUiSlider.set(valueElement.value); //сброс насыщенности эффекта //resetImage
+  // imagePreview.style.transform = '';
+  // imagePreview.style.filter = ''; //возврат к эффекту Оригинал
+  // imagePreview.className = 'effects__preview--none';
+  // sliderElement.classList.add('hidden'); //скрываем ползунок
 };
 
 const setUserFormSubmit = (onSuccess) => {
@@ -125,9 +141,11 @@ const setUserFormSubmit = (onSuccess) => {
         () => {
           onSuccess();
           unblockSubmitButton();
+          showSuccessMessage();
+          resetForm();
         },
         () => {
-          showAlert('Не удалось отправить форму. Попробуйте ещё раз');
+          showErrorMessage();
           unblockSubmitButton();
         },
         new FormData(evt.target),
