+++ js/message.js
@@ -1,21 +1,4 @@
-
 import {isEscapeKey} from './util.js';
-import {setUserFormSubmit} from './forms.js';
-import {sendData} from './api.js';
-
-const sliderElement = document.querySelector('.effect-level__slider');
-const valueElement = document.querySelector('.effect-level__value');
-valueElement.value = 100; //сброс уровня насыщенности
-const scaleInput = document.querySelector('input[name="scale"]');
-const imagePreview = document.querySelector('.img-upload__preview img');
-const loadingForm = document.querySelector('.img-upload__form');
-
-const resetForm = () => {
-  loadingForm.reset(); //сброс полей формы
-  sliderElement.noUiSlider.set(valueElement.value); //сброс насыщенности эффекта
-  scaleInput.value = '100%'; //возврат к масштабу 100%
-  imagePreview.style.filter = ''; //возврат к эффекту Оригинал
-};
 
 const successMessageTemplate = document
   .querySelector('#success')
@@ -53,7 +36,7 @@ const showSuccessMessage = () => {
 const showErrorMessage = () => {
   const errorMessageElement = errorMessageTemplate.cloneNode(true);
   document.addEventListener('keydown', onMessageEscdown);
-  const errorButton = errorMessageElement.querySelector('error__button');
+  const errorButton = errorMessageElement.querySelector('.error__button');
   errorButton.addEventListener('click', onErrorButtonClick);
   bodyElement.append(errorMessageElement);
   bodyElement.style.overflow = 'hidden';
@@ -67,11 +50,4 @@ function hideMessage () {
   bodyElement.style.overflow = 'auto';
 }
 
-const onSendDataSuccess = () => {
-  resetForm();
-  showSuccessMessage();
-};
-
-setUserFormSubmit(async (data) => {
-  await sendData(onSendDataSuccess, showErrorMessage, data);
-});
+export {showSuccessMessage, showErrorMessage};
